neuronode-backend/scripts/setup/requirements-dev.txt
# Development Dependencies für Neuronode
# Diese Pakete werden nur für die Entwicklung benötigt

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.24.0  # Für API-Tests

# Code Quality & Linting
black>=23.7.0  # Code-Formatierung
isort>=5.12.0  # Import-Sortierung
flake8>=6.0.0  # Linting
mypy>=1.5.0    # Type Checking

# Development Tools
ipython>=8.14.0     # Bessere REPL
jupyter>=1.0.0      # Notebooks für Experimente
notebook>=7.0.0     # Jupyter Notebook Interface

# Documentation
sphinx>=7.1.0       # Dokumentation generieren
sphinx-rtd-theme>=1.3.0  # ReadTheDocs Theme

# Debugging & Profiling
pdb++>=0.10.3       # Besserer Debugger
memory-profiler>=0.61.0  # Memory Profiling
line-profiler>=4.1.0     # Line-by-line Profiling

# Development Server
watchdog>=3.0.0     # File watching für Hot Reload
python-dotenv>=1.0.0  # .env file loading (falls nicht in main requirements)

# API Development
httpie>=3.2.0       # HTTP client für API-Tests
requests-mock>=1.11.0  # Mock HTTP requests

# Database Development
neo4j-driver[test]>=5.11.0  # Neo4j test utilities

# Type Stubs
types-requests>=2.31.0
types-PyYAML>=6.0.0

# Performance Testing
locust>=2.16.0      # Load testing
